Analysis: Use ES version of PatternAnalyzer for prebuilt version which is not susceptible to default locale issues.
