+++ src/main/java/org/elasticsearch/index/analysis/PatternAnalyzer.java
@@ -1,0 +1,58 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.index.analysis;
+
+import org.apache.lucene.analysis.Analyzer;
+import org.apache.lucene.analysis.TokenStream;
+import org.apache.lucene.analysis.core.LowerCaseFilter;
+import org.apache.lucene.analysis.core.StopFilter;
+import org.apache.lucene.analysis.pattern.PatternTokenizer;
+import org.apache.lucene.analysis.util.CharArraySet;
+
+import java.io.Reader;
+import java.util.regex.Pattern;
+
+/**
+* Replacement for Lucene's PatternAnalyzer which is not susceptible to lowercasing localization issues.
+*/
+public final class PatternAnalyzer extends Analyzer {
+    private final org.apache.lucene.util.Version version;
+    private final Pattern pattern;
+    private final boolean lowercase;
+    private final CharArraySet stopWords;
+
+    public PatternAnalyzer(org.apache.lucene.util.Version version, Pattern pattern, boolean lowercase, CharArraySet stopWords) {
+        this.version = version;
+        this.pattern = pattern;
+        this.lowercase = lowercase;
+        this.stopWords = stopWords;
+    }
+
+    @Override
+    protected TokenStreamComponents createComponents(String s, Reader reader) {
+        final TokenStreamComponents source = new TokenStreamComponents(new PatternTokenizer(reader, pattern, -1));
+        TokenStream result = null;
+        if (lowercase) {
+             result = new LowerCaseFilter(version, source.getTokenStream());
+        }
+        result = new StopFilter(version, (result == null) ? source.getTokenStream() : result, stopWords);
+        return new TokenStreamComponents(source.getTokenizer(), result);
+    }
+}
